# Conflicts:
#	build.gradle
#	src/main/java/twilightforest/entity/ai/goal/ThrowRiderGoal.java
#	src/main/java/twilightforest/entity/boss/Naga.java
#	src/main/java/twilightforest/item/PeacockFanItem.java
#	src/main/java/twilightforest/network/ThrowPlayerPacket.java
#	src/main/resources/data/twilightforest/recipes/patchouli_logbook.json
#	src/main/resources/data/twilightforest/recipes/patchouli_logbook_alt.json

+++ src/main/java/twilightforest/entity/ai/goal/ThrowRiderGoal.java
@@ -93,7 +93,7 @@ public class ThrowRiderGoal extends MeleeAttackGoal {
 			}
 
 			if (rider instanceof ServerPlayer player) {
-				ThrowPlayerPacket message = new ThrowPlayerPacket((float) throwVec.x(), (float) throwVec.y(), (float) throwVec.z());
+				ThrowPlayerPacket message = new ThrowPlayerPacket(throwVec.x(), throwVec.y(), throwVec.z());
 				TFPacketHandler.CHANNEL.sendToClient(message, player);
 			}
 		}
